Move scene change detection into ContextInner
It will need access to these fields for lookahead scene detection
